feat(F-450): replace global-error.tsx with error.tsx
