+++ frontend/src/interfaces/ErrorPageProps.ts
@@ -1,0 +1,4 @@
+export interface ErrorPageProps {
+  error: Error & { digest?: string };
+  reset: () => void;
+}
